Refactor: Extract the JSON namespace to a separate header
